Split up the ungainly PerformImplementation.hpp.
